pyproject.toml
[tool.poetry]
name = "com_redis_session"
version = "0.1.0"
description = "High level API to redis as session storage backend"
authors = [
    "Danila Korobkov <korobkov.danila.yurevich@gmail.com>",
]
maintainers = [
    "Danila Korobkov <korobkov.danila.yurevich@gmail.com>",
]
readme = "README.md"
homepage = "https://github.com/DanilaKorobkov/com_redis_session"
repository = "https://github.com/DanilaKorobkov/com_redis_session"
documentation = "https://github.com/DanilaKorobkov/com_redis_session/browse/README.md"

[tool.poetry.dependencies]
python = "^3.8"
aioredis = "^1.3"
attrs = "^21.2"

[tool.poetry.dev-dependencies]
add-trailing-comma = "^2.1"
autopep8 = "^1.5"
bandit = "^1.7"
flake8 = "^3.9"
isort = "^5.8"
mypy = "^0.812"
pylint = "^2.8"
pytest = "^6.2"
pytest-cov = "^2.12"
pytest-sugar = "^0.9"
marshmallow = "^3.12"
factory-boy = "^3.2"
ujson = "^4.0"
com_redis_test_utils = { branch="develop", git = "git@github.com:DanilaKorobkov/com_redis_test_utils.git" }

[build-system]
requires = ["poetry-core>=1.0"]
build-backend = "poetry.core.masonry.api"
